cobalt/pod: drop xnpod_stop_thread() service
This routine was aimed at implementing non-destructive thread
termination services (e.g. arinc653's STOP_SELF()), which can and
should be done from userland instead. Besides, xnpod_stop_thread()
blurred the semantics of XNDORMANT in a confusing way.

As a consequence of this change, XNDORMANT has a single meaning: it
tags a thread which is yet to be started via xnpod_start_thread(), and
therefore only happens once in any thread's lifetime.