cobalt/thread, posix: clearly distinguish exit vs finalization handlers
The exit handler is called locklessly with interrupts enabled, for
dropping all resources allocated by the personality, except the TCB
which may still be referenced for accessing resources maintained by
the Cobalt core.

The finalization handler shall drop the TCB and may therefore assume
that no further reference to this memory may be done afterwards from
the Cobalt core.