cobalt/pod: fix deletion of dormant threads
A dormant thread which receives a cancellation request through
xnpod_cancel_thread() must wake up first, for reaching a cancellation
point asap.

In no way shall this thread's resources be dropped over the caller's
context, this would break the basic assumption with the Cobalt core,
that resources must be claimed on behalf of the exiting thread only.