cobalt/pod: drop PEXEC and FATAL condition bits
XNPEXEC makes no sense anymore as it used to protect our kernel
handlers from running when no real-time services are available yet.

It turns out that the Cobalt core is now initialized early, and those
handlers are not even installed until the interrupt pipeline is
requested to forward the trap and syscall events, which only
happens...after the system has been initialized.

Finally, exposing the fatal bit as a global state flag, only to test
it within the panic handler was quite overkill.