cobalt/thread: move finalizer call out of any critical section
There is no point in calling the thread finalization handler while
holding a lock, since it is supposed to release the TCB to some
allocator, which may incur some work.

Besides, the thread has entered an innocuous state long ago before the
finalizer is eventually called, so there is nothing to protect in its
TCB at this point.

Finalizers should do proper locking locally if need be.