lib/cobalt: extend control over library init procedure
__cobalt_defer_init is a weak boolean variable which may be overriden
by any external compilation unit for preventing the library
constructor from running the initialization code, including binding
the process to the real-time system.

In deferred init mode, __libcobalt_init() may be called later for
running the real-time init code, which will NOT include the
auto-shadowing procedure for the calling thread though.

CAUTION: upon fork(), the parent's setting for init deferral is
inherited by the child process. This means that a forkee shall call
__libcobalt_init() for binding to the real-time system, if forked from
a parent which previosuly asked for a deferred init.

In addition, __cobalt_main_prio is a weak integer value which may be
overriden by any external compilation unit for selecting the
scheduling policy and priority, to be set when auto-shadowing the main
thread. For this reason, this setting only applies when
__cobalt_defer_init is unset.

__cobalt_main_prio == 0, forces [SCHED_OTHER, 0]
__cobalt_main_prio > 0, forces  [SCHED_FIFO, __cobalt_main_prio]
__cobalt_main_prio < 0, reuses the current scheduling parameters.