copperplate/threadobj: introduce __threadobj_set_scheduler()
This is an internal helper aimed at tracking changes to the thread
priority, without performing any other related fixups. We use this in
POSIX-based APIs to resync the copperplate thread state with external
updates done directly via the pthread_setschedparam() service.

This helper should disappear when we actually implement a copperplate
service to set any arbitrary policy and scheduling parameters, in
replacement of threadobj_set_priority() - which is logically broken
and uselessly restrictive.