lib/cobalt, cobalt/shadow: allow user-space to issue serial debug output
In desperate cases, it may be convenient to write raw debug output
directly to the console UART from the current domain, without any
intermediate buffering.

This internal debug service is available from the inner Cobalt
interface as __cobalt_serial_debug(const char *fmt, ...). The
formatted output is directed to __ipipe_serial_debug().

NOTE: for the output to take place, the support for serial console and
CONFIG_IPIPE_DEBUG must be enabled in the kernel.