cobalt/signal: implement thread group sending
When a signal is sent via the kill() interface and the destination
thread is not waiting for that signal, we try to deliver it to any of
the threads which belong to the same process. If no thread is waiting
for this signal, it is queued for the destination thread as usual.

This change does not affect the behavior of pthread_kill(), which
remains strictly thread-targeted.