cobalt/posix/signal: Account for dequeued timer signals
If a timer is deleted, we remove potential pending events from the
signal queue of the receiver. But we do not (and cannot) clear
sigpending in this case. So account for the case the sigpending still
contains a signal event for which no queue element is present: fix up
sigpending and recheck the signal status.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>