cobalt/posix/thread: introduce ->thread_setsched() extension handler
We need a way to inform extensions about a policy change to happen
soon for a thread.

The new ->thread_setsched() handler is called with interrupts off,
nklocked held, right before the change takes place. The handler is
passed the new scheduling class and parameters which will be applied
upon return.

The handler is guaranteed that the change will take place upon return
(no error path behind the extension call).