cobalt/rtdm: introduce rtdm_task_should_stop()
As a consequence of rebasing Xenomai kernel threads over linux
kthreads, the Cobalt core may not guarantee synchronous deletion of
kernel threads anymore.

We introduce rtdm_task_should_stop(), as a mean for RTDM driver tasks
to poll for a pending termination request, which might have been sent
to them by a call to rtdm_task_destroy(). If a true value is returned,
the caller should wrap up and exit gracefully.

In addition, to match the expectations of the existing RTDM user base
with respect to the rtdm_task_destroy() semantics, the latter now
explicitly waits for the deleted thread to terminate, before returning
to the caller.