cobalt/thread: introduce mode bit for disabling lock breaking
This commit introduces XNTRAPLB, a new thread mode bit for disabling
the scheduler lock breaking feature.

In the default case, a thread which holds the scheduler lock is
allowed to drop it temporarily for sleeping. If this mode bit is set,
such thread would return immediately with XNBREAK set from
xnthread_suspend().

The consequence of disabling the lock breaking feature may range from
-EINTR received by the offending API service, to entering an infinite
loop if the userland code restarts the syscall upon -EINTR. In the
latter case, the watchdog should trigger.