powerpc: sanitize SPE handling
switchtest: we support SPE in user-space, but there is no point in
checking for fprs sanity between context switches since we have no
such registers in this case.

kernel handling: enabling SPE implicitly means disabling HW_FPU
handling in the Xenomai nucleus, so tracking the SPE context for
real-time kthreads is unsupported. However, we do have proper SPE
save/restore code in our user context switching code, which enables it
for real-time threads in userland.

This patch is based on:
http://lists.debian.org/debian-powerpc/2011/07/msg00026.html