posix, vrtx: catch unexpected sync errors with thread trampoline
If __real_sem_wait() fails synchronizing with the trampoline, we
simply can't continue, as this usually reveals a general issue with
regular IPCs on the platform.

This also plugs a nasty bug causing the POSIX thread trampoline to
execute random start code, as the argument structure living on its
parent thread's stack could be trashed, once the latter has returned
to the caller because of a synchronization failure.