cobalt/sched: drop sched flag debugging feature
This feature was aimed at detecting changes in the scheduler state
which were not covered by the XNSCHED flag. This situation would be
interpreted as some code affecting the run queue outside of the legit
routines.

As a pick_next() handler may reassess the need for switching context
based on information specific or even internal to the policy module,
XNSCHED may not always reflect the latest scheduler state change.