copperplate/threadobj: dismantle the TCB gradually on failure
We may have to drop the TCB resources for threads which failed to
start internally. To this end, we need to split the uninit/cleanup
handlers, the same way we have the converse init/setup handlers.

At this chance, threadobj_destroy() is renamed threadobj_uninit(), in
order to match the common naming scheme (e.g. syncobj_destroy() vs
synchobj_uninit()).

This fixes issue:
http://www.xenomai.org/pipermail/xenomai/2014-January/029966.html