copperplate/registry: introduce per-file private data, downgrade to FUSE_2.5 API
In order to make use of the private fuse_file_info::fh data passed to
the fs handlers, we have to work around a glibc dynamic linker bug (*):
https://sourceware.org/bugzilla/show_bug.cgi?id=12977
http://www.xenomai.org/pipermail/xenomai/2013-October/029301.html

We do this by downgrading to the fuse API rev 2.5, which binds
references to versioned fuse symbols explicitly to the proper
implementation for FUSE_2.5, working around the symbol lookup in dl.

(*) otherwise, the dl bug causes a reference to fuse_main() to be
resolved as a call to fuse_mail_real_compat22() instead of
fuse_mail_real_compat25(), introducing a nasty mismatch with the
fuse_file_info layout.