alchemy: Replace static variable no_alchemy_task with macro
The current definition of a static const variable representing an
invalid alchemy task is both C++-incompatible and may leave variables in
the .bss of modules behind that are including task.h, even if they don't
use the symbol. So replace it with a macro that builds the required
struct on-the-fly.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>