copperplate/internal: add parent-child handshake protocol at spawn
We want to provide the ability to run a prologue code on behalf of the
emerging child context, which the parent shall wait for completion of
before continuing.

This feature can be used to guarantee that a child thread has
performed a set of actions before the caller of
copperplate_create_thread() is allowed to refer to it.