copperplate/registry: reap sysregd selectively
We want to allow application code to wait for children exits
explicitly and selectively using wait*() calls, while preventing a
failing sysregd to move to the zombie state. Therefore, bluntly
leaving the SIGCHLD disposition to SIG_IGN upon return from this
routine is not an option.

To solve this issue, first we ignore SIGCHLD to plug a potential race
while forking the daemon, then we trap it to a valid handler
afterwards, once we know the daemon pid. This handler will selectively
reap the registry daemon, and only this process, leaving all options
open to the application code for reaping its own children as it sees
fit.