cobalt/shadow: propagate CPU migration event to personalities
This patch extends the personality operations, with the .move_thread()
event handler, which is called when the regular kernel decides to move
a thread to another CPU.

The handler is called nklock held, IRQs off. The Cobalt TCB of the
affected thread is not updated yet, when the call is made.