boilerplate/lock: prevent FMR in safe unlock op
Since dropping the lock may result in the memory holding the restored
cancel state to become stale, we'd better read it prior to unlocking.

Typically, this case would happen with call sites like
threadobj_unlock(), fetching the cancel state word from the TCB of a
thread which pends on its own lock, prior to finalizing. If that
thread preempts threadobj_unlock() (as a result of the lock drop) then
exits, the cancel state may become stale before threadobj_unlock()
refers to it.