cobalt/kernel: convert legacy trace markers to kernel tracepoints
Rebase trace statements in the Cobalt core over the generic tracepoint
API. A few former trace markers which did not help in analysing the
dynamic behavior of the system have been dropped in the process.

This work is originally based on Jan Kiszka's trace events conversion
patch series for Xenomai 2.6.