copperplate/notifier: sanitize and simplify - again
Those changes build on the fact that the notification support is there
exclusively for implementing the thread suspend/resume mechanism over
Mercury. Therefore we may drop any code which is not directly aimed at
supporting this feature.

Assuming this, locking may now be handled directly from the call
sites, all located in the threadobj implementation.

As a bonus, this change set also fixes the issue discussed there:
http://www.xenomai.org/pipermail/xenomai/2014-April/030804.html