copperplate/notifier: signal suspended thread directly
The notifier will remain dedicated to supporting thread-directed
suspend/resume operations eventually, and there is no gain in issuing
a signaled message compared to solely signaling the target thread.

Therefore we drop the pipe formerly used to trigger a fasync
notification on the read side, replacing it by direct signaling of the
target thread.

This patch is a slightly updated version of Matthias Schneider's
proposal:
http://www.xenomai.org/pipermail/xenomai/2014-May/030822.html