doc/doxygen: simplify and sanitize build rules
We keep two generated targets only, folding all APIs into a single
main document:

- html generation for the online manual
- PDF generation for the manual book

Both Cobalt and Mercury cores are documented in the same document,
core-specific APIs are few, as well as variations for common APIs.

The TOC is utterly broken currently, due to contents which moved
around and wrong markup here and there. This will be addressed next.