doc: add convenience rules for producing specific output
From the top Makefile:

$ make doc/doxygen
$ make doc/man
$ make doc/asciidoc

Respectively recurse in the doc/{doxygen, man, asciidoc} subdirs,
invoking the default rule there.

In addition:

$ make doc/html
$ make doc/pdf

Recurse in the doc/{doxygen, man, asciidoc} subdirs, invoking the html
or pdf rule there. In doc/doxygen/, this specifically rebuilds the
online manual in HTML form, or as a PDF book.