doc: clarify terms used for prebuilt documentation
Rename doc/generated to doc/prebuilt, to convey the idea that such
information might not be up-to-date, until it is eventually
regenerated.

At this chance, add the doc/resync rule (aliased to doc/prebuild) to
the top Makefile, for bootstrapping the rebuilding process. This
replaces the former generate-doc rule from the doc/ sub-tree.