doc: fold man pages into the asciidoc tree
The asciidoc processing recipe only differs marginally for all
document types, so there is no point in duplicating the hairy stuff
only to separate the source files.

Besides, we don't have that many asciidoc source files, and that
number may even decrease as we ramp up online contents with the
upcoming web site, so it is unlikely that doc/asciidoc will ever get
crowded.