lib/smokey: introduce infrastructure for building smoke tests
A new copperplate-based API called "smokey" is introduced, which
automates cumbersome boilerplate tasks when writing smoke tests.
Smokey provides a simple plugin-oriented infrastructure, with the
ability to host and run a set of tests embodied into a single
executable.

A companion program is provided in testsuite/smokey, for running the
individual tests formerly available under testsuite/unit, illustrating
the usage of the smokey API.  testsuite/regression will be converted
next.