alchemy: do not publish objects to clusters until they are fully built
We must NOT push Alchemy objects to clusters before they have been
fully built, including the registration phase.

Otherwise, a high priority task waiting on the binding service for the
object to appear in a cluster might preempt the creation call
immediately when the latter is updated, then destroy the half-baked
resource being pushed, leading to invalid memory references and all
sorts of unpleasant outcomes.