cobalt/rtdm, drivers: move .open, .socket handlers to fd_ops
With the recent introduction of a generic file operation descriptor
which drops the legacy rt/nrt forms of these handlers, there is no
reason to keep them floating alone in the rtdm_device definition.

This patch moves the .open and .socket handlers into the rtdm_fd_ops
descriptor with other I/O handlers, fixing up the in-tree drivers
accordingly.