cobalt/rtdm: merge with Cobalt/POSIX personality
There is no need for a distinct RTDM personality. Features which can
be accessed from user-space belong to the POSIX I/O API, while RTDM
tasks directly sit on top of the core personality.

This patch merges the RTDM syscalls into the Cobalt/POSIX table, and
dropping the RTDM personality in the same move.