cobalt/kernel: use single switch for internal consistency checks
Now that we have a compact core system, having most of the internal
consistency checks enabled via a single switch is developer-friendly -
i.e. so that we don't inadvertently miss a disabled assertion -
provided this keeps the overhead low.

This patch merges the former NUCLEUS and RTDM debug levels into the
COBALT one.