cobalt/syscall: split syscall handlers and modes
x32 will require a sparse syscall table, with modes shared between 64
and 32 bit call forms, so it makes sense to maintain handlers and
modes in separate tables.

In addition, although we don't benefit from loading the handler and
the call mode from the same cacheline anymore, the generated assembly
in 64bit mode is much, much better, due to the natural alignment of
the call table entry.