cobalt/uapi, copperplate, alchemy, trank: make cobalt_event 32/64bit-neutral
Using long integers for representing event bit masks potentially
creates a portability issue for applications between 32 and 64bit CPU
architectures. This issue is solved by using 32bit integers on 32/64
bit machines, which is normally more than enough for encoding the set
of events received by a single event group.

In addition, this enables the sharing of the event value between a
64bit kernel support and 32bit libraries without thunks.