ksrc: fix dereference of dangling pointer
When enabling CONFIG_DEBUG_PAGEALLOC and CONFIG_IPIPE_TRACER, a reference to a
destroyed task is detected at __xnpod_schedule() entry, when it is trying to
trace the PID of the currently running task.

This is because __xnpod_schedule() uses the root tcb user_task pointer, which
may be, in fact, invalid at this point, the root tcb user_task pointer is only
valid when in primary mode (after a call to xnpod_leave_root()).

To avoid this issue, we introduce xnthread_current_user_task() and
xnthread_current_user_pid() which will return the current thread and its pid
without relying on the tcb user_task pointer.