arm: fix compilation issue with --enable-debug
When passing --enable-debug and no CFLAGS, the user-space support is
compiled without optimization. Unfortunately, on ARM, when compiling
in thumb2 mode with frame pointers, this causes compilation errors
when inline assembly uses the "r7" register, which is also the frame
pointer. We can not very well stop using the r7 register since it is,
for instance, part of the user/kernel syscall ABI, so when we detect
this situation, we add the -fomit-frame-pointer flag. This will make
the resulting code a little less easy to debug, but at least it will
compile.