cobalt/heap: drop support for heap extension
In-kernel users only maintain fixed-size heaps, so there is no upside
in making the latter extendable, but only useless overhead. This patch
drops the support for the "extent" abstraction, assuming a single
contiguous storage area for any given heap.

In addition, the page map meta-data is moved away from the
user-defined storage area, so that the client code may actually
receive the full amount of memory it gave us originally (barring the
external fragmentation).

The page map is allocated internally instead via kmalloc(), which adds
the secondary_mode_only() requirement for both xnheap_init() and
xnheap_destroy(). All existing callers are fine with this already.