cobalt/heap: limit all sizes and offsets to 32bit values
The allocator cannot handle more than 2Gb of storage area by design,
and the uapi/ bits already assume that sizes and offsets produced by
Cobalt heaps fit in a 32bit word, for 32/64 ABI neutrality.

Therefore we may turn all relevant long types to u32 in the
implementation, to clearly state the addressing limit of this
allocator.