cobalt/rtdm, cobalt/posix/process: extend mmap() support to MMU-less
A file mapping on no-MMU requires the backing device to advertise
direct mapping capabilities (i.e. BDI_CAP_MAP_DIRECT), which a file
hooked on the anon inode doesn't.

To fix this, we issue our internal mapping requests (i.e. for memdev
heaps and mayday page) on a file pointer to /dev/mem instead, on
MMU-mess systems (only).

In addition, the .get_unmapped_area handler is added to the RTDM file
descriptor operations, since this will be a prerequisite for allowing
drivers to honor mapping requests to (quasi-)memory devices on
MMU-less the same way.