cobalt/rtdm: drop poll_delay from rtdm_dev_unregister()
Sleep indefinitely on a plain waitqueue instead of polling, until the
device becomes idle. For this reason, the routine loses its return
value as well, as drivers are most often not willing to deal with
receiving a device busy condition from a module exit routine (which is
the place devices should be unregistered from).

Drivers which really want to deal with such condition should simply
use module refcounting in their own code.