cobalt/rtdm/fd: assume current context is target for all fd ops
Except in the only weird case of forcing a close on any random
kernel-originated connection through the procfs interface
(open_fildes), but this case is on its way out.

For the sake of mental sanity, we really want to assume that only the
current thread may affect its own file table, which is what this patch
enforces.

The rtdm_fd cleanup thread is a non-issue, as we are not supposed to
run any fd ops recursively on behalf of a ->close() handler anyway.

This change also fixes a rampant inconsistency where a Cobalt user
thread running some driver code would actually look up into the wrong
RTDM file table when using the rt_dev_* interface
(i.e. __xnsys_global_ppd would not be the right context in such a
case).