cobalt/timer: anticipate on the expiry date
The gravity is directly substracted from the timer expiry date,
instead of applying gravity on the timer shot date and/or within the
timer processing loop. This prevents the overhead that may be caused
by the need for rescheduling early shots.

However, this does NOT prevent early shots in case the clock gravity
value for the target context is over-pessimistic.