smokey/sched-quota: throttle test threads to stop calibration
It is not totally clear why we need that, but as a matter of fact the
test control thread appears to be stuck in glibc's pthread_cancel() on
SMP/x86 during the first couple of runs while calibrating, despite it
has higher priority than the threads to cancel.

There is some suspicion on the adverse effect of running the
glibc-internal pthread_cancel_init() code at the first invocation of
pthread_cancel(), but no formal proof. It behaves like the test
control thread waits for some event completion in pthread_cancel() -
even before sys_tkill is issued - which never happens, as the test
threads spinning on the same CPU tend to lock it out.

Bottom-line is that something looks fishy here, and this change is
more a workaround than a fix.