cobalt/uapi: allow up to 128 distinct clock identifiers
As more POSIX clocks are being defined by the regular kernel, we need
to reserve a larger namespace (32) for their identifiers. At this
chance, the Cobalt namespace for static and dynamic clocks is extended
too (32 and 64 ids respectively).

As a consequence of this change, we limit the pshared attribute to a
plain boolean value to get additional encoding space in some base type
(i.e. condvar attribute).