rtnet: rtskb_pool changes (2)
Move the per-driver rtskb_pool to rtnet_devices. At this chance, rework
rtdev_reference/rtdev_dereference to lock/unlock module loading, also lock
the module when an interface is up, so that an interface must be down in order
to allow removing it.

A new call rtnetdev_alloc_rtskb is added to be used instead of dev_alloc_rtskb
to allocate devices from the device pool, its first argument being a pointer to
a device instead of a pointer to a pool.