cobalt/intr: fix first enable
In order to enable an irq after having requested it, upper layers
use xnintr_enable() after xnintr_attach().

But xnintr_enable() does not call ipipe_enable_irq if the XN_IRQSTAT_DISABLED
is not set in the intr object status field.

So, in particular, xnintr_enable() called after xnintr_attach() is a nop,
which results in RT driver irqs lines not being enabled, and the driver not
receiving irqs.

Fix this by setting the XN_IRQSTAT_DISABLED bit in the intr object status
field in xnintr_init().