copperplate/registry: Switch to /var/run as default mountpoint root
Creating directories in a user/admin-owned and typically persistent
filesystem for the purpose of mounting volatile filesystems on top is
a suboptimal default. Today /run should be used for such purposes, but
older distros may only expose the current FHS standard /var/run. So use
that instead.

This comes with the advantage that we will now typically work against
tmpfs, thus won't stumble if / is read-only.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>