cobalt/pipe: Avoid running init_waitqueue_head under xnlock
This operation can be more complex than simple structure initialization,
namely if lockdep is enabled.

The approach take here is to break up user connection into two stages.
The first one is "lingering connection", which simply reserves the pipe
for the caller of xnpipe_open. After that we release xnlock, perform the
Linux part of the initialization and then finish the second stage under
xnlock again to reach USER_CONN just as before.

This issue can be triggered by running smokey test #10 (xddp).

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>