lib/cobalt: Fix ordering of fork handler
Currently, fork handlers run in more or less random order after
successive forks because pthread_atfork is re-executed for some
handlers. This can cause the heap being unmapped after __init_cobalt was
executed, leaving cobalt_current_window in an invalid state. One result
of this is a crashing sysregd when run as unprivileged user, thus when
it forked fusermount from the forked daemon.

Fix the issue by performing the after-fork work centrally, in a
well-defined order.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>